Exposes TokenValidator func and JWTTokenValidator interface, updates
This brings in the change from one of the use cases we have for this library. It's been a part of the vendored version only, which obviously breaks easily and can't be tracked normally, so should be a part of the repo.
Also updates usage of go-jose as it changed Json* type names to all uppercase some time ago, which breaks things unless the application pins a specific version (which needs to be deduced).